Fix handling of case where non-serialized backend loads missing key
AFAIK, beaker will try to prevent this by taking out a read lock, checking for the key's existance, and then reading it. But if the read lock isn't implemented for a backend (or isn't cross host), the item could be deleted between the check and the get and then the decode will blow up. This prevents that.